Move shared geo funcs to boundaries/helpers.py
